chore: simplify windows_subsystem attribute
