+++ examples/api/src-tauri/Cargo.toml
@@ -6,6 +6,9 @@ edition = "2021"
 rust-version = "1.59"
 license = "Apache-2.0 OR MIT"
 
+[lib]
+name = "api_lib"
+
 [build-dependencies]
 tauri-build = { path = "../../../core/tauri-build", features = ["codegen", "isolation"] }
 
@@ -32,7 +35,6 @@ features = [
 ]
 
 [target."cfg(target_os = \"windows\")".dependencies]
-window-vibrancy = "0.2"
 window-shadows= "0.2"
 
 [target.'cfg(any(target_os = "android", target_os = "ios"))'.dependencies]
