+++ examples/api/src-tauri/src/main.rs
@@ -1,12 +1,7 @@
-#![cfg_attr(
-  all(not(debug_assertions), target_os = "windows"),
-  windows_subsystem = "windows"
-)]
-
-#[cfg(desktop)]
-mod desktop;
+// Prevents additional console window on Windows in release, DO NOT REMOVE!!
+#![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
 fn main() {
   #[cfg(desktop)]
-  desktop::main();
+  api_lib::run();
 }
