+++ examples/tauri-dynamic-lib/src-tauri/src/lib.rs
@@ -5,10 +5,7 @@
 // This is an example of a tauri app built into a dll
 // Calling lib_test1 within the dll will launch the webview
 
-#![cfg_attr(
-  all(not(debug_assertions), target_os = "windows"),
-  windows_subsystem = "windows"
-)]
+#![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
 #[no_mangle]
 pub extern "C" fn run_tauri() {
