+++ examples/web/core/tauri/src/main.rs
@@ -1,7 +1,4 @@
-#![cfg_attr(
-  all(not(debug_assertions), target_os = "windows"),
-  windows_subsystem = "windows"
-)]
+#![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
 #[tauri::command]
 fn greet(window: tauri::Window, name: String) {
