+++ tooling/cli/templates/plugin/backend/examples/vanilla/src-tauri/src/main.rs
@@ -1,7 +1,5 @@
-#![cfg_attr(
-  all(not(debug_assertions), target_os = "windows"),
-  windows_subsystem = "windows"
-)]
+// Prevents additional console window on Windows in release, DO NOT REMOVE!!
+#![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
 fn main() {
   tauri::Builder::default()
